Fixed a bug (not a bug actually, but a pseudo-feautre, which I now think is not possible to implement but will only cause trouble) reported by Abhishek(i.abhi27@gmail.com): The basic problem was that linkpit allowed linking of URLs like 	google.com 	www.google.com Because of which when these URL were redirected by Location: header, would cause redirection to http://linkpit.co.cc/google.com because the browser thought that it was relatively linked, just as Location: abc.gif would tell browser to fetch http://site.com/abc.gif Now because of this, Linkpit again thought that it should push this URL to the database, hence repeating the whole cycle.
This bug is now fixed. URLs must now begin with http://or ftp:// or something something starting with :// so as to make it a proper URL(see functions.php).
As suggested by long ago duckinator at ##programming, it would be better to do a routine checkup on all URLs which are dead and destroy them to save space.
We have to do that.

As again I think we need to document what a propper tag and propper URL is.

Some other changes;
(*) Added message about the [tag] notation
(*) Added README