Search Info Pop Up Box in Autocompletion.
A pop up dialog box is displayed when an option
is selected from the result pop up menu. At the
moment the dialog shows TAG NAME, URL and HITS
COUNTER.

To obtain data related to the TAG for the pop
up, AJAX request is made to search-info.php
file. This file is temporary and will be
integrated with other search services in
single file.